third_party/blink/web_tests/platform/generic/http/tests/navigation/ping-cookie-expected.txt

PingLoader dispatched to 'http://127.0.0.1:8000/navigation/resources/save-Ping.php?test=cookie'.
Ping sent successfully
CONTENT_TYPE: text/ping
HTTP_COOKIE: hello=world
HTTP_PING_FROM: http://127.0.0.1:8000/navigation/ping-cookie.html
HTTP_PING_TO: http://127.0.0.1:8000/navigation/resources/check-ping.php?test=cookie
REQUEST_METHOD: POST
